requirements.txt
absl-py==2.1.0
altair==5.4.1
annotated-types==0.7.0
astunparse==1.6.3
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
click==8.1.7
distro==1.9.0
firebase-admin==6.5.0
flatbuffers==24.3.25
fsspec==2024.9.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.2
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h5py==3.12.1
httplib2==0.22.0
huggingface-hub==0.25.1
jiter==0.5.0
keras==3.5.0
libclang==18.1.1
loguru==0.7.2
Markdown==3.7
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
narwhals==1.8.4
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==1.50.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.12.1
pandas==2.2.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
PyJWT==2.9.0
pyparsing==3.1.4
PySocks==1.7.1
pytz==2024.2
rclone==0.4.4
regex==2024.9.11
rsa==4.9
safetensors==0.4.5
sentencepiece==0.2.0
smmap==5.0.1
streamlit==1.38.0
sympy==1.13.3
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
tokenizers==0.20.0
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.1
triton==3.0.0
tzdata==2024.2
uritemplate==4.1.1
watchdog==4.0.2
Werkzeug==3.0.4
wrapt==1.16.0
